+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -1,0 +1,4 @@
+package ru.yandex.practicum.filmorate.storage;
+
+public class InMemoryUserStorage {
+}
